+++ src/main/java/entities/Admin.java
@@ -1,18 +1,8 @@
 package entities;
 
-public class Admin extends Account implements AdminFactory {
+public class Admin extends Account {
     public Admin (String password, String accountName) {
         super(password, accountName);
-    }
-
-    @Override
-    public boolean giveWarning() {
-        return false;
-    }
-
-    @Override
-    public boolean restrictUser() {
-        return false;
     }
 
     public void addMessage(String message) { System.out.println(message); }
