--- src/main/java/entities/AdminFactory.java
@@ -1,6 +1,0 @@
-package entities;
-
-public interface AdminFactory {
-    public boolean restrictUser();
-    public boolean giveWarning();
-}
