+++ src/main/java/entities/DatabaseConnect1.java
@@ -1,0 +1,7 @@
+package entities;
+
+public class DatabaseConnect1 {
+    public User findRandomAdmin() {
+
+    }
+}
