+++ src/main/java/use_cases/record_review_use_case/RecordReportInputBoundary.java
@@ -1,4 +1,4 @@
-package record_review_use_case;
+package use_cases.record_review_use_case;
 
 public interface RecordReportInputBoundary {
     RecordReportOutputData createReport(RecordReportInputData inputData);
