+++ src/main/java/use_cases/record_review_use_case/RecordReportOutputData.java
@@ -1,4 +1,4 @@
-package record_review_use_case;
+package use_cases.record_review_use_case;
 
 public class RecordReportOutputData {
     private String timestamp;
