+++ src/main/java/use_cases/record_review_use_case/RecordReportPresenter.java
@@ -1,4 +1,4 @@
-package record_review_use_case;
+package use_cases.record_review_use_case;
 import UI.RecordReportResultViewModel;
 
 import java.text.SimpleDateFormat;
@@ -7,12 +7,23 @@ public class RecordReportPresenter implements RecordReportOutputBoundary {
     RecordReportOutputData recordReportOD;
     RecordReportResultViewModel recordReportRVM;
 
+    /**
+     * Constructor for the presenter class.
+     * @param recordReportOD    An instance of the output data.
+     * @param recordReportRVM   An instance of the view model.
+     */
     public RecordReportPresenter(RecordReportOutputData recordReportOD,
                                  RecordReportResultViewModel recordReportRVM) {
         this.recordReportOD = recordReportOD;
         this.recordReportRVM = recordReportRVM;
     }
 
+    /**
+     * Calls the success view of the use case view model if success is true and the fail view
+     * of the it if success is false. Returns the instance of the recordReportOutputBoundary.
+     * @param success This is the success boolean variable.
+     * @return recordReportOD This is the instance of the recordReportOutputBoundary.
+     */
     @Override
     public RecordReportOutputData responseView(boolean success) {
         if (success) recordReportRVM.successView();
@@ -21,6 +32,9 @@ public class RecordReportPresenter implements RecordReportOutputBoundary {
         return recordReportOD;
     }
 
+    /**
+     * Assigns a timestamp to the instance of the recordReportOutputBoundary.
+     */
     public void setTimestamp() {
         String timestamp = new SimpleDateFormat("yyyy.MM.dd.HH.mm.ss").format(new java.util.Date());
         recordReportOD.setTimestamp(timestamp);
