+++ src/main/java/use_cases/restrict_user_use_case/RestrictUserController.java
@@ -1,4 +1,4 @@
-package restrict_user_use_case;
+package use_cases.restrict_user_use_case;
 
 public class RestrictUserController {
     RestrictUserInputBoundary restrictUserIB;
@@ -7,7 +7,9 @@ public class RestrictUserController {
         this.restrictUserIB = restrictUserIB;
     }
 
-    //The reported user will often be the user restricted, hence they use the same controller / interactor / presenter.
+    // The reported user will often be the user restricted, hence they use the same
+    // controller, interactor, and presenter.
+
     public RestrictUserOutputData accessReport(String reportedUserID) {
         RestrictUserInputData inputData = new RestrictUserInputData(reportedUserID);
         return restrictUserIB.accessReport(inputData);
