+++ src/main/java/use_cases/restrict_user_use_case/RestrictUserOutputBoundary.java
@@ -1,4 +1,4 @@
-package restrict_user_use_case;
+package use_cases.restrict_user_use_case;
 
 public interface RestrictUserOutputBoundary {
     RestrictUserOutputData displayReport(RestrictUserOutputData outputData);
