+++ src/main/java/use_cases/restrict_user_use_case/RestrictUserOutputData.java
@@ -1,4 +1,4 @@
-package restrict_user_use_case;
+package use_cases.restrict_user_use_case;
 
 import java.util.ArrayList;
